feat/refactor: title in header and maximum length of list name
